+++ src/main/java/com/sealog/backend/domain/feature/auth/service/AuthService.java
@@ -7,15 +7,6 @@ import com.sealog.backend.global.exception.CustomException;
 public interface AuthService {
 
     /**
-     * 회원가입
-     * @param request 회원가입 요청 DTO
-     * @return 생성된 사용자 엔티티 (컨트롤러에서 쿠키 설정)
-     * @throws CustomException 이메일이 이미 존재하는 경우 (CONFLICT)
-     * @throws CustomException 닉네임이 이미 존재하는 경우 (CONFLICT)
-     */
-    User signUp(AuthRequest.SignUpRequest request);
-
-    /**
      * 로그인
      * @param request 로그인 요청 DTO
      * @return 인증된 사용자 엔티티 (컨트롤러에서 쿠키 설정)
