+++ src/main/java/com/sealog/backend/domain/feature/auth/service/AuthServiceImpl.java
@@ -22,32 +22,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class AuthServiceImpl implements AuthService {
 
     private final UserRepository userRepository;
-    private final UserFileRepository userFileRepository;
-    private final FileMetadataRepository fileMetadataRepository;
     private final PasswordEncoder passwordEncoder;
-    private final FileStorageService fileStorageService;
-    private final UserValidatorService userValidatorService;
 
-    @Override
-    @Transactional
-    public User signUp(AuthRequest.SignUpRequest request) {
-        // 이메일 중복 검사
-        userValidatorService.validateDuplicateEmail(request.getEmail());
 
-        // 닉네임 중복 검사
-        userValidatorService.validateDuplicateNickname(request.getNickname());
-
-        // 비밀번호 암호화 및 User 생성
-        User user = User.builder()
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .name(request.getName())
-                .nickname(request.getNickname())
-                .role(UserRole.USER)
-                .build();
-
-        return userRepository.save(user);
-    }
 
     @Override
     public User login(AuthRequest.LoginRequest request) {
