+++ src/main/java/com/sealog/backend/domain/feature/user/service/AdminUserServiceImpl.java
@@ -1,0 +1,44 @@
+package com.sealog.backend.domain.feature.user.service;
+
+import com.sealog.backend.domain.feature.auth.dto.AuthRequest;
+import com.sealog.backend.domain.feature.user.entity.User;
+import com.sealog.backend.domain.feature.user.entity.UserRole;
+import com.sealog.backend.domain.feature.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class AdminUserServiceImpl implements AdminUserService {
+
+    private final UserValidatorService userValidatorService;
+    private final PasswordEncoder passwordEncoder;
+    private final UserRepository userRepository;
+
+    @Override
+    @Transactional
+    public User signUp(AuthRequest.SignUpRequest request) {
+        // 이메일 중복 검사
+        userValidatorService.validateDuplicateEmail(request.getEmail());
+
+        // 닉네임 중복 검사
+        userValidatorService.validateDuplicateNickname(request.getNickname());
+
+        // 비밀번호 암호화 및 User 생성
+        User user = User.builder()
+                .email(request.getEmail())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .name(request.getName())
+                .nickname(request.getNickname())
+                .role(UserRole.USER)
+                .build();
+
+        return userRepository.save(user);
+    }
+}
